Completed 100%

+++ manage.php
@@ -12,7 +12,8 @@ session_start();
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
 </head>
 <body>
-   <div class="container-fluid bg-light mt-5">
+   <div class="container-fluid bg-light mt-5 py-5">
+    <h1 class="text-center text-primary">User DashBoard</h1>
                 <!-- for delete message -->
                 <?php
              if(isset($_GET['msg'])){
@@ -25,19 +26,7 @@ session_start();
               }
             }
             ?>
-            <!-- For login success message -->
-            <?php
-             if(isset($_GET['msg'])){
-              $delete_msg = $_GET['msg'];
-              if($delete_msg =='loginsuccess'){
-                ?>
-                <div class="alert alert-success" role="alert">
-                  <h4 class="alert-heading text-center">Login successfully</h4>
-                </div><?php
-              }
-            }
-            ?>
-   <div class="row">
+   <div class="row py-5">
     <div class="col-12 col-sm-6 col-lg-6">
         <div class="card text-center bg-light">
             <img class="card-img-top m-auto" src="./assests/uploads/user.png" alt="Card image cap" style="width:150px;height:150px">
@@ -47,6 +36,7 @@ session_start();
         </div>
     </div>
     <div class="col-12 col-sm-6 col-lg-6">
+    <a href="createtask.php" style="float:right"><button class="btn btn-primary btn-lg">Add Task</button></a>
         <table class="table w-100">
             <thead>
               <tr>
@@ -58,31 +48,34 @@ session_start();
             </thead>
               <tbody>
                 <?php
-                $query = "SELECT * FROM task";
+                $uid = $_SESSION['id'];
+                $query = "SELECT * FROM task WHERE userid= $uid";
                 $query_result = mysqli_query($conn,$query);
-                $i=0;
-                while($task = mysqli_fetch_array($query_result)){
-                  $i++;
-                  ?>
-                  <tr>
-                  <th scope="row"><?php echo $i ?></th>
-                  <td><?php echo $task['title']?></td>
-                  <td><?php echo $task['description']?></td>
-                  <td><a href="edit.php?id=<?php echo $task['id'];?>">
-                  <button class="btn btn-primary mx-4">Edit</button>
-                  </a>
-                  <a href="delete.php?id=<?php echo $task['id'];?>">
-                  <button class="btn btn-danger mx-4">Delete</button>
-                  </a>
-                </td>
-                </tr><?php
+                $count = mysqli_num_rows($query_result);
+                if($count == 1){
+                  $i=0;
+                  while($task = mysqli_fetch_array($query_result)){
+                    $i++;
+                    ?>
+                    <tr>
+                    <th scope="row"><?php echo $i ?></th>
+                    <td><?php echo $task['title']?></td>
+                    <td><?php echo $task['description']?></td>
+                    <td><a href="edit.php?id=<?php echo $task['id'];?>">
+                    <button class="btn btn-primary mx-4">Edit</button>
+                    </a>
+                    <a href="delete.php?id=<?php echo $task['id'];?>">
+                    <button class="btn btn-danger mx-4">Delete</button>
+                    </a>
+                  </td>
+                  </tr><?php
+                  }
                 }
                 ?>
             </tbody>
         </table>
     </div>
   </div>
-  <div class="alert alert-primary text-center" role="alert"><a href="createtask.php"><button class="btn btn-primary btn-lg">Add Task</button></a></div>
 </div>
 </body>
 </html>
